Fixed compile-frontend

+++ compile-frontend.sh
@@ -5,9 +5,14 @@ if [ "$1" != "develop" -a "$1" != "master" ]; then
     exit
 fi
 
+if [ "$2" = "" ]; then
+    echo "Incorrect container name"
+    exit
+fi
+
 cd resources/frontend
 git checkout $1 && git pull origin $1
-cd ../../docker && docker-compose exec workspace bash -c '
+cd ../../docker && docker exec $2 bash -c '
     rm -rf public/build
     cd resources/frontend
     rm -rf node_modules
